+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,35 @@
 
 ### Patch Changes
 
+- Updated dependencies [e9897eb]
+  - ai@4.1.58
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [092fdaa]
+  - ai@4.1.57
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [80be82b]
+- Updated dependencies [8109a24]
+  - ai@4.1.56
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [1531959]
+  - ai@4.1.55
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [ee1c787]
   - ai@4.1.54